+++ lib/views/play/play_view.dart
@@ -15,6 +15,7 @@ export 'providers/play_time_provider.dart';
 export 'providers/playing_difficulty_provider.dart';
 export 'providers/win_status_provider.dart';
 
+export 'widgets/board.dart';
 export 'widgets/board_square.dart';
 export 'widgets/done_dialog.dart';
 export 'widgets/mines_left_counter.dart';
